Ai_ResumeAnalyser/requirements.txt
absl-py==1.4.0
altair==5.2.0
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.3.4
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==1.6.0
Babel==2.12.1
backcall==0.2.0
backports.functools-lru-cache==2.0.0
beautifulsoup4==4.11.1
black==22.1.0
bleach==4.1.0
blinker==1.7.0
blis==0.7.10
cachetools==4.2.2
catalogue==2.0.9
certifi==2020.12.5
cffi==1.15.0
channels==4.0.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.7
cloudinary==1.25.0
colorama==0.4.6
comm==0.1.2
comtypes==1.1.14
confection==0.1.1
constantly==15.1.0
cryptography==41.0.5
cycler==0.11.0
cymem==2.0.7
daphne==3.0.2
dateparser==1.1.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-appconf==1.0.4
django-binary-database-files==1.0.15
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-googledrive-storage==1.5.2
django-on-heroku==1.1.2
djangorestframework==3.12.4
docutils==0.19
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
entrypoints==0.4
executing==0.8.2
fastjsonschema==2.18.0
Flask==3.0.3
flatbuffers==23.1.21
fonttools==4.29.1
fqdn==1.5.1
future==0.18.3
gast==0.4.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.40
google==3.0.0
google-api-core==1.27.0
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
GoogleNews==1.6.5
googletrans==3.0.0
greenlet==3.0.3
grpcio==1.51.3
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httplib2==0.19.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.0.0
incremental==22.10.0
indic-transliteration==2.3.57
install==1.3.5
ipykernel==6.21.2
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.3
joblib==1.1.0
json5==0.9.6
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.2.0
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.24.0
kaldi-io==0.9.5
keras==2.11.0
keyboard==0.13.5
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kiwisolver==1.3.2
langcodes==3.3.0
libclang==15.0.6.1
lxml==5.2.2
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==3.0.1
MouseInfo==0.1.3
murmurhash==1.0.9
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.4
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.22.2
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
overrides==7.4.0
packaging==20.9
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathy==0.10.2
pdfminer.six==20231228
pefile==2022.5.30
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.5.0
preshed==3.0.8
prometheus-client==0.13.1
prompt-toolkit==3.0.37
protobuf==3.17.0
psutil==5.9.4
psycopg2-binary==2.8.6
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyKaldi==0.0.1
PyMsgBox==1.0.9
pynput==1.7.6
pyOpenSSL==23.3.0
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.1
python-docx==1.1.2
python-json-logger==2.0.7
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.0
QtPy==2.3.0
redis==4.4.0
referencing==0.30.2
regex==2022.10.31
requests==2.25.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.0
roman==4.1
rpds-py==0.9.2
rsa==4.7.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.12.2
Send2Trash==1.8.2
sentencepiece==0.1.99
service-identity==23.1.0
shellingham==1.5.4
six==1.16.0
smart-open==6.3.0
smmap==5.0.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.9.0
spotipy==2.22.0
SQLAlchemy==2.0.29
sqlparse==0.4.1
srsly==2.4.7
srt==3.5.2
stack-data==0.1.4
streamlit==1.29.0
swig==4.1.1
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.13.1
testpath==0.5.0
thinc==8.1.12
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
transliterate==1.10.2
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typer==0.12.3
typing-extensions==3.7.4.3
tzdata==2022.1
tzlocal==4.2
uri-template==1.3.0
uritemplate==3.0.1
urllib3==1.26.4
validators==0.22.0
vosk==0.3.45
waitress==2.0.0
wasabi==0.10.1
watchdog==3.0.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.4
Werkzeug==3.0.2
whitenoise==5.2.0
widgetsnbextension==4.0.5
wikipedia==1.4.0
wrapt==1.14.1
youtube-search-python==1.6.6
zipp==3.11.0
zope.interface==6.1